+++ lib/generators/rails_performance/install/templates/initializer.rb
@@ -8,11 +8,11 @@ if defined?(RailsPerformance)
 
     # Recent Requests configuration
     config.recent_requests_time_window = 60.minutes
-    # confog.recent_requests_limit = nil # numnber of recent requests
+    # config.recent_requests_limit = nil # number of recent requests
 
     # Slow Requests configuration
     config.slow_requests_time_window = 4.hours
-    # config.slow_requests_limit = 500 # numnber of slow requests
+    # config.slow_requests_limit = 500 # number of slow requests
     config.slow_requests_threshold = 500 # ms
 
     config.debug = false # currently not used>
@@ -52,5 +52,9 @@ if defined?(RailsPerformance)
     config.skipable_rake_tasks = ["webpacker:compile"]
     config.include_rake_tasks = false
     config.include_custom_events = true
+
+    # If enabled, the system monitor will be displayed on the dashboard
+    # to enabled add required gems (see README)
+    # config.system_monitor_duration = 24.hours
   end
 end