+++ test/dummy/app/controllers/home_controller.rb
@@ -2,7 +2,10 @@ class HomeController < ApplicationController
   def index
     rand(10).times { AdvancedWorker.perform_async }
     SecondWorker.perform_async
-    sleep(rand(1000) * 0.001)
+
+    100000.times { rand(1000000000) / rand(1000232323.9) }
+
+    sleep(rand(500) * 0.001)
     10.times { User.where(first_name: "X#{rand(100)}").count }
     RailsPerformance.measure "index", "home controller" do
       sleep(0.1)
@@ -11,7 +14,7 @@ class HomeController < ApplicationController
   end
 
   def about
-    sleep(rand(1000) * 0.001)
+    sleep(rand(500) * 0.001)
     rand(10).times { SimpleWorker.perform_async }
     10.times { User.where(first_name: "X#{rand(100)}").count }
 
@@ -19,7 +22,7 @@ class HomeController < ApplicationController
   end
 
   def contact
-    sleep(rand(1000) * 0.001)
+    sleep(rand(500) * 0.001)
     rand(5).times { SimpleWorker.perform_async }
     rand(3).times { AdvancedWorker.perform_async }
     10.times { User.where(first_name: "X#{rand(100)}").count }