+++ CHANGES.md
@@ -1,5 +1,14 @@
 - Unreleased
 
+- 1.4.1
+  - switch to ApexCharts https://github.com/igorkasyanchuk/rails_performance/pull/117
+  - upgraded dummy app to rails 8 https://github.com/igorkasyanchuk/rails_performance/pull/119
+
+- 1.4.0
+  - track CPU, memory, storage: https://github.com/igorkasyanchuk/rails_performance/pull/111
+  - use UTC time: https://github.com/igorkasyanchuk/rails_performance/pull/114
+  - custom expiration time for system monitoring report: https://github.com/igorkasyanchuk/rails_performance/pull/115/files
+
 - 1.3.3
   - little improvements and bug fixes
   - changes in the readme